Refactor service/datasource for search & pagination
- Remove BookRepository and have Service use DataSource directly
- Add search param to GetBooks and implement GetBooksCount in GORM datasource
- Populate paginated response (PaginatedBooks) in Service with TotalItems (int64) and TotalPages
- Change PaginatedBooks.TotalItems type to int64
- Use BASE_URL (with default) for cover URLs in gormsql
- Move embedded static file serving into gin router setup
- Update router module dependencies

+++ dto.go
@@ -4,7 +4,7 @@ type PaginatedBooks struct {
 	Data       []Book `json:"data"`
 	Page       int    `json:"page"`
 	PageSize   int    `json:"page_size"`
-	TotalItems int    `json:"total_items"`
+	TotalItems int64  `json:"total_items"`
 	TotalPages int    `json:"total_pages"`
 }
